itest: move multi hop claim tests into one file
